Para fins de padronização, os tipos Long foram alterados para Integer.

+++ SistemaLobinho/src/main/java/engtelecom/bcd/model/AreasDeConhecimento.java
@@ -13,7 +13,7 @@ import lombok.RequiredArgsConstructor;
 public class AreasDeConhecimento {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idAreaDeConhecimento;
+    private Integer idAreaDeConhecimento;
 
     @NonNull
     @Column(nullable = false)
